feat: Add getMonthlyTotalAmount API endpoint
This commit adds a new API endpoint called getMonthlyTotalAmount to the BillController. This endpoint calculates the total amount of bills for the current month and returns the result as a JSON response. It uses the SQLite database to query the bill table and retrieve the sum of the amount column for the bills created between the first day and the last day of the current month.